Moves endpoint code out of app/lib into app
