+++ app/moco-routing/config/startCampaignTransitionsConfigModel.js
