+++ app/moco-routing/config/yesNoPathsConfigModel.js
