+++ config/smsConfigsLoader.js
@@ -11,11 +11,11 @@ app.ConfigName = {
 
 var conn = require('./connectionConfig');
 var configModelArray = [
-  rootRequire('app/lib/chatbot/models/donorsChooseBotModel')(conn),
-  rootRequire('app/lib/chatbot/models/donorsChooseMocoCampaignModel')(conn),
-  rootRequire('app/lib/ds-routing/config/startCampaignTransitionsConfigModel')(conn),
-  rootRequire('app/lib/ds-routing/config/yesNoPathsConfigModel')(conn),
-  rootRequire('app/lib/reportback/reportbackConfigModel')(conn),
+  rootRequire('app/chatbot/models/donorsChooseBotModel')(conn),
+  rootRequire('app/chatbot/models/donorsChooseMocoCampaignModel')(conn),
+  rootRequire('app/moco-routing/config/startCampaignTransitionsConfigModel')(conn),
+  rootRequire('app/moco-routing/config/yesNoPathsConfigModel')(conn),
+  rootRequire('app/reportback/reportbackConfigModel')(conn),
 ];
 
 var logger = rootRequire('lib/logger');
