+++ test/BullyText2014.js
@@ -2,11 +2,11 @@ var assert = require('assert')
   , express = require('express')
   , emitter = rootRequire('app/eventEmitter')
   , connectionOperations = rootRequire('config/connectionOperations')
-  , gameMappingModel = rootRequire('app/lib/sms-games/models/sgGameMapping')(connectionOperations)
-  , gameModel = rootRequire('app/lib/sms-games/models/sgCompetitiveStory')(connectionOperations)
-  , userModel = rootRequire('app/lib/sms-games/models/sgUser')(connectionOperations)
-  , SGCompetitiveStoryController = rootRequire('app/lib/sms-games/controllers/SGCompetitiveStoryController')
-  , smsHelper = rootRequire('app/lib/smsHelpers')
+  , gameMappingModel = rootRequire('app/sms-games/models/sgGameMapping')(connectionOperations)
+  , gameModel = rootRequire('app/sms-games/models/sgCompetitiveStory')(connectionOperations)
+  , userModel = rootRequire('app/sms-games/models/sgUser')(connectionOperations)
+  , SGCompetitiveStoryController = rootRequire('app/sms-games/controllers/SGCompetitiveStoryController')
+  , smsHelper = rootRequire('app/smsHelpers')
   , testHelper = require('./testHelperFunctions')
   ;
