+++ test/testHelperFunctions.js
@@ -2,9 +2,9 @@ var assert = require('assert')
   , express = require('express')
   , emitter = require('../app/eventEmitter')
   , connectionOperations = rootRequire('config/connectionOperations')
-  , gameMappingModel = rootRequire('app/lib/sms-games/models/sgGameMapping')(connectionOperations)
-  , gameModel = rootRequire('app/lib/sms-games/models/sgCompetitiveStory')(connectionOperations)
-  , userModel = rootRequire('app/lib/sms-games/models/sgUser')(connectionOperations)
+  , gameMappingModel = rootRequire('app/sms-games/models/sgGameMapping')(connectionOperations)
+  , gameModel = rootRequire('app/sms-games/models/sgCompetitiveStory')(connectionOperations)
+  , userModel = rootRequire('app/sms-games/models/sgUser')(connectionOperations)
   , SGCompetitiveStoryController = require('../app/lib/sms-games/controllers/SGCompetitiveStoryController')
   , smsHelper = require('../app/lib/smsHelpers')
   , _ = require('underscore')
